src/app/auth/guards/auth.guard.ts
import { Injectable } from '@angular/core';
import { CanActivate, UrlTree, Router } from '@angular/router';
import { Observable } from 'rxjs';
import { DEFAULT_ROUTE_SUCCESS } from '../interfaces/token-response';
import { AuthService } from '../services/auth.service';

@Injectable({
  providedIn: 'root'
})
export class AuthGuard implements CanActivate {
  constructor(private authService: AuthService, private router: Router) {}
  canActivate(): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
      if (this.authService.isLoggedIn()) {
        this.router.navigate([DEFAULT_ROUTE_SUCCESS]);
      }
    return !this.authService.isLoggedIn();
  }
  
}
